requirements.txt
# Core Framework & RTC
livekit>=1.0.25          # Das neue 1.0+ SDK
livekit-agents>=1.4.1    # Wichtig: Beinhaltet die neuen Session-Logiken
livekit-api>=1.0.7       # Passende Server-API

# Plugins (Gemini & Co)
livekit-plugins-google>=1.4.1      # Für Gemini 2.0 Realtime
livekit-plugins-elevenlabs>=1.3.11 # Aktuellste ElevenLabs Integration
livekit-plugins-silero>=1.3.0      # VAD Plugin

# Audio & Math
numpy>=2.4.2             # Neuestes NumPy
sounddevice>=0.5.5       # Für lokale Wiedergabe

# HTTP Client
httpx>=0.27.0            # Für Calendly API Calls

# Utilities
python-dotenv>=1.2.1     #
asyncio                  # (Standard-Bibliothek, meist nicht in req.txt nötig, aber ok)